% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{c1}
\alias{c1}
\title{Limit distribution: Original CUSUM, one-sided alternative}
\usage{
c1(w, npts, gamma)
}
\arguments{
\item{w}{Sample Weiner process}

\item{npts}{Number of points in discretization of Weiner process}

\item{gamma}{tuning parameter}
}
\value{
numeric
}
\description{
Limit distribution: Original CUSUM, one-sided alternative
}
