% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptFor-class.R
\name{calcCusum,cptFor-method}
\alias{calcCusum,cptFor-method}
\title{Sequential changepoint analysis of Forecast Errors}
\usage{
\S4method{calcCusum}{cptFor}(x, gamma = 0, rootDir = "./")
}
\arguments{
\item{gamma}{tuning parameter used in CUSUM detector}

\item{cptFor}{An object of \code{cptFor} with the following slots filled:
\itemize{
\item{errors: }{forecast errors must be provided}
\item{m: }{length of training period must be provided}
}}
}
\value{
an object of class \code{cptFor}
}
\description{
Takes an object of class \code{cptFor} and (re)caclulates the CUSUM of the forecast errors and squared forecast errors.
}
\examples{
obj = new("cptFor", errors=stats::rnorm(500), m=300)
obj = calCusum(obj)
summary(obj)
plot(obj)
}
