% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptFor-class.R
\docType{class}
\name{cptFor-class}
\alias{cptFor-class}
\title{An S4 class for a forecast changepoint object}
\value{
An object of class \code{cptFor}.
}
\description{
An S4 class for a forecast changepoint object
}
\section{Slots}{

\describe{
\item{\code{X}}{ANY. Raw time series}

\item{\code{m}}{numeric. Length of training data}

\item{\code{model}}{ANY. Summary of model used for forecasting}

\item{\code{errors}}{ANY. Forecast errors, one-step-ahead}

\item{\code{errorsVar}}{numeric. Variance of forecast errors in training data}

\item{\code{errors2Var}}{numeric. Variance of squares of forecast errors}

\item{\code{cusum}}{numeric. CUSUM values of forecast errors}

\item{\code{cusum2}}{numeric. CUSUM values of squares of forecast errors}

\item{\code{cusumThreshold}}{numeric. threshold values for CUSUM values of forecast errors}

\item{\code{cusum2Threshold}}{numeric. threshold values for CUSUM values of squares of forecast errors}

\item{\code{tau}}{numeric. Time point when changepoint is flagged based upon forecast errors}

\item{\code{tau2}}{numeric. Time point when changepoint is flagged based upon squares of forecast errors}
}}

