% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{cptFor-methods}
\alias{cptFor-methods}
\alias{summary,cptFor-method}
\alias{show,cptFor-method}
\alias{plot,cptFor-method}
\title{cptFor Methods}
\usage{
\S4method{summary}{cptFor}(object)

\S4method{show}{cptFor}(object)

\S4method{plot}{cptFor}(x)
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{cptFor}}}

\item{x}{An object of S4 class \code{\linkS4class{cptFor}}}
}
\description{
Methods for objects with S4 class \code{\linkS4class{cptFor}}.
}
\section{Functions}{
\itemize{
\item \code{summary,cptFor-method}: Summary of \code{cptFor} object.

This includes the version of the package used; date the analysis was performed; the
changepoint detector used; and information on whether a change was detected or not.

\item \code{show,cptFor-method}: Show \code{cptFor} object.

This includes the \code{summary} method alongside all the slots of the \code{cptFor} object.

\item \code{plot,cptFor-method}: Plot of \code{cptFor} object.

This produces a plot of the forecast errors; CUSUM detector(s) and indicates whether the threshold
was reached and a changepoint detected. This is a \code{\link[ggplot2]{ggplot}} object and thus
can be stored and manipulated as required.
}}

\examples{
ans = cptForecast(c(stats::rnorm(400), stats::rnorm(100,3)), m=300)
summary(ans)
show(ans)
plot(ans)
}
