% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptSeqCUSUM.R
\name{cptSeqCUSUM}
\alias{cptSeqCUSUM}
\title{Sequential Changepoint Analysis via a CUSUM Detector}
\usage{
cptSeqCUSUM(
  X,
  m = ceiling(0.5 * length(X)),
  detector = "PageCUSUM",
  gamma = 0,
  sigma2 = NULL,
  critValue = "Lookup",
  alpha = 0.05,
  samples = 1000,
  npts = 500,
  Class = TRUE
)
}
\arguments{
\item{X}{time series. Can be a numeric vector or a \code{ts} object}

\item{m}{length of training period}

\item{detector}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}

\item{gamma}{tuning parameter in the weight function. See details below}

\item{sigma2}{variance of data in training period. Default is \code{NULL} and variance is estimated using \code{var()} (assumes indpendence).}

\item{critValue}{numeric or character:
\itemize{
\item{numeric: }{This will be used as the critical value. Note this is not the raw threshold.}
\item{'Lookup': }{a pre-simulated critical value is used if available.}
\item{'Simulate': }{a critical value is simulated, dependent on scenario this can be time consuming.}
}}

\item{alpha}{asymptotic type 1 error rate. Not used if critValue is a numeric.}

\item{samples}{Only used if \code{critValue='Simulate'}. Number of sample of Weiner processes used to calculate critical value.}

\item{npts}{Only used if \code{critValue='Simulate'}. Number of points in each sample of a Weiner process used to calculate critical value.}

\item{Class}{Logical. Return an object of class \code{cptFor}.}
}
\value{
an object of class \code{cptFor}.
If \code{Class=FALSE} a list with
\itemize{
\item tau: Time a change is flagged
\item cusum: Vector of cusum values
\item threshold: Vector of threshold values
\item sigma2: Variance of points in training period
}
}
\description{
This function performs sequential changepoint analysis using a chosen CUSUM detector.
}
\examples{
X = c(stats::rnorm(400), stats::rnorm(100, 2))
ans = cptSeqCUSUM(X, m=300)
summary(ans)
}
