% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptSeqCUSUM.R
\name{cptSeqCUSUMerroCheck}
\alias{cptSeqCUSUMerroCheck}
\title{Error checking - cptSeqCUSUM}
\usage{
cptSeqCUSUMerroCheck(
  X = X,
  m = m,
  detector = detector,
  gamma = gamma,
  sigma2 = sigma2,
  critValue = critValue,
  alpha = alpha,
  samples = samples,
  npts = npts,
  Class = Class
)
}
\arguments{
\item{X}{time series. Can be a numeric vector or a \code{ts} object}

\item{m}{length of training period}

\item{detector}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}

\item{gamma}{tuning parameter in the weight function. See details below}

\item{sigma2}{variance of data in training period. Default is \code{NULL} and variance is estimated using \code{var()} (assumes independence).}

\item{critValue}{numeric or character:
\itemize{
\item{numeric: }{This will be used as the critical value. Note this is not the raw threshold.}
\item{'Lookup': }{a pre-simulated critical value is used if available.}
\item{'Simulate': }{a critical value is simulated, dependent on scenario this can be time consuming.}
}}

\item{alpha}{asymptotic type 1 error rate. Not used if critValue is a numeric.}

\item{samples}{Only used if \code{critValue='Simulate'}. Number of sample of Weiner processes used to calculate critical value.}

\item{npts}{Only used if \code{critValue='Simulate'}. Number of points in each sample of a Weiner process used to calculate critical value.}

\item{Class}{Logical. Return an object of class \code{cptFor}.}
}
\description{
Performs error checking on arguments given to cptSeqCUSUM
}
