% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critValLookup.R
\name{critValLookup}
\alias{critValLookup}
\title{Lookup Critical Value}
\usage{
critValLookup(gamma = 0, alpha = 0.05, CUSUMtype = "Page", oneSidedAlt = FALSE)
}
\arguments{
\item{gamma}{tuning parameter}

\item{alpha}{type 1 error}

\item{CUSUMtype}{character. Choice between 'Page' and 'Original' CUSUM detector. See details below}

\item{oneSidedAlt}{Logical. Use of the one-sided or two-sided CUSUM detectors}
}
\value{

}
\description{
Lookup pre-simulated critical value of asymptotic distribution of chosen CUSUM detector
}
\examples{
critValLookup(gamma=0,
              alpha=0.05,
              CUSUMtype='Page',
              oneSidedAlt=FALSE)
}
