% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumGenerator.R
\name{cusumGenerator}
\alias{cusumGenerator}
\title{Creates a function for calculating CUSUM statistic}
\usage{
cusumGenerator(X, m, oneSidedAlt = FALSE)
}
\arguments{
\item{X}{Time series}

\item{m}{Length of training period}

\item{oneSidedAlt}{Logical. Testing the one-sided or two-sided alternative}
}
\value{
A function with one parameter $k$.
}
\description{
Creates a function which returns the CUSUM statistic at a given time point $k$.
}
