% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusumPageGenerator.R
\name{cusumPageGenerator}
\alias{cusumPageGenerator}
\title{Creates a function for calculating Page's CUSUM statistic}
\usage{
cusumPageGenerator(X, m, oneSidedAlt = FALSE)
}
\arguments{
\item{X}{Time series}

\item{m}{Length of training period}

\item{oneSidedAlt}{Logical. Testing the one-sided or two-sided alternative}
}
\value{
A function with one parameter $k$
}
\description{
Creates a function which returns Page's CUSUM statistic at a given time point $k$.
}
