% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastAutoArima.R
\name{forecastAutoArima}
\alias{forecastAutoArima}
\title{Forecast with Auto ARIMA}
\usage{
forecastAutoArima(X, m, trend = FALSE, Class = TRUE, ...)
}
\arguments{
\item{X}{time series. Can be numeric vector or a \code{ts} object}

\item{m}{length of training period}

\item{trend}{Logical. Include a deterministic linear trend.}

\item{Class}{Logical. Return an object of class \code{cptFor}.}

\item{...}{Additional parameters to pass to \code{auto.arima}.}
}
\value{
object of class \code{cpt.for}. If \code{Class=FALSE} a list with elements:
\itemize{
\item{model: }{fitted ARIMA model based upon training data}
\item{errors: }{forecast errors}
}
}
\description{
Calculates one-step-ahead forecast errors from an automatically chosen ARIMA model based upon training data
}
\examples{
X = stats::arima.sim(list(ar=0.3), n=500)
ans = forecastAutoArima(X, m=300)
summary(ans)
}
