% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastEs.R
\name{forecastEs}
\alias{forecastEs}
\title{Forecast with Exponential Smoothing}
\usage{
forecastEs(X, m, Model = "XXX", Class = TRUE, ...)
}
\arguments{
\item{X}{time series. Can be numeric vector or a \code{ts} object}

\item{m}{length of training period}

\item{Model}{Type of Exponential Smoothing models to allow.}

\item{Class}{Logical. Return an object of class \code{cptFor}.}

\item{...}{Additional parameters to pass to \code{es}.}
}
\value{
object of class \code{cptFor}. If \code{Class=FALSE} a list with elements:
\itemize{
\item{model: }{fitted Exponential Smoothing model based upon training data}
\item{errors: }{forecast errors}
}
}
\description{
Calculate one-step-ahead forecast errors from an automatically chosen Exponential Smoothing model based upon training data.
}
\examples{
#' X = stats::arima.sim(list(ar=c(0.3, 0.1, -0.4), ma=c(0.5, 0.2)), n=500)
ans = forecastEs(X, m=300)
summary(ans)
}
