% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastMlp.R
\name{forecastMlp}
\alias{forecastMlp}
\title{Forecast with ML}
\usage{
forecastMlp(X, m, trainData, trend = FALSE, Class = TRUE, ...)
}
\arguments{
\item{X}{time series. Can be numeric vector or a \code{ts} object}

\item{m}{numeric. Length of stable forecasts}

\item{trainData}{training data. This data is used to train the ML model}

\item{trend}{Logical. Include a deterministic linear trend.}

\item{Class}{Logical. Return an object of class \code{cptFor}.}

\item{...}{Additional parameters to pass to \code{mlp}.}
}
\value{
object of class \code{cpt.for}. If \code{Class=FALSE} a list with elements:
\itemize{
\item{model: }{fitted ML model based upon training data}
\item{errors: }{forecast errors}
}
}
\description{
Calculates one-step-ahead forecast errors from a machine learning model based upon training data
}
\examples{
\dontrun{
X = stats::arima.sim(list(ar=0.3), n=500)
trainData = stats::arima.sim(list(ar=0.3), n=100)
ans = forecastMlp(X, m=100, trainData=trainData)
summary(ans)
}
}
