% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{limDistGenerator}
\alias{limDistGenerator}
\title{Samples from Limit Distributions of CUSUM Detectors}
\usage{
limDistGenerator(
  CUSUMtype = "Page",
  oneSidedAlt = FALSE,
  gamma = 0,
  samples = 1000,
  npts = 500,
  progressBar = TRUE
)
}
\arguments{
\item{CUSUMtype}{character. Choice between 'Page' and 'Original' CUSUM detector. See details below}

\item{oneSidedAlt}{Logical. Use of the one-sided or two-sided CUSUM detectors}

\item{gamma}{tuning parameter in the weight function}

\item{samples}{Number of samples to simulate from asymptotic distribution}

\item{npts}{Number of points in discretization of Weiner process}

\item{progressBar}{Logical. Show progress bar}

\item{alpha}{Type 1 error}
}
\value{
numeric vector of samples
}
\description{
Generates required number of samples from the asymptotic limit distribution of the chosen CUSUM detector under the scenario of no change.
}
