% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{p1}
\alias{p1}
\title{Limit distribution: Page CUSUM, one-sided alternative}
\usage{
p1(w, npts, gamma)
}
\arguments{
\item{w}{Sample Weiner process}

\item{npts}{Number of points in discretization of Weiner process}

\item{gamma}{tuning parameter}
}
\value{
numeric
}
\description{
Limit distribution: Page CUSUM, one-sided alternative
}
