% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{p1inf}
\alias{p1inf}
\title{Minimum of Weiner process in Page's one-sided CUSUM limit distribution}
\usage{
p1inf(w, t, npts)
}
\arguments{
\item{w}{Sample Weiner process}

\item{t}{Max point to find minimum within}

\item{npts}{Number of points in discretization of Weiner process}
}
\value{
numeric
}
\description{
Minimum of Weiner process in Page's one-sided CUSUM limit distribution
}
