% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{p2}
\alias{p2}
\title{Limit distribution: Page's CUSUM, two-sided alternative}
\usage{
p2(w, npts, gamma)
}
\arguments{
\item{w}{Sample Weiner process}

\item{npts}{Number of points in discretization of Weiner process}

\item{gamma}{tuning parameter}
}
\value{
numeric
}
\description{
Limit distribution: Page's CUSUM, two-sided alternative
}
