% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{p2sup}
\alias{p2sup}
\title{Maximum of Weiner process in Page's two-sided CUSUM limit distribution}
\usage{
p2sup(w, t, npts, gamma)
}
\arguments{
\item{w}{Sample Weiner process}

\item{t}{Max point to find minimum within}

\item{npts}{Number of points in discretization of Weiner process}

\item{gamma}{tuning parameter}
}
\value{
numeric
}
\description{
Maximum of Weiner process in Page's two-sided CUSUM limit distribution
}
