% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptFor-class.R
\name{plot,cptFor,ANY-method}
\alias{plot,cptFor,ANY-method}
\title{Plot of \code{cptFor} object}
\usage{
\S4method{plot}{cptFor,ANY}(x)
}
\arguments{
\item{cptFor}{an object of class \code{cptFor}}
}
\value{
ggplot
}
\description{
Produces a ggplot containg the raw data, forecast errors, CUSUM of forecast errors and squared forecast errors along with thresholds.
}
