% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptFor-class.R
\name{plotMulti,list-method}
\alias{plotMulti,list-method}
\title{Plot of list of \code{cptFor} objects}
\usage{
\S4method{plotMulti}{list}(x, names)
}
\arguments{
\item{x}{a list of \code{cptFor} objects}
}
\value{
ggplot
}
\description{
Produces a ggplot where each column is the plot of each \code{cptFor} object in the list.
}
