% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMulti.R
\name{plotMulti}
\alias{plotMulti}
\title{Plot a list of \code{cptFor} objects}
\usage{
plotMulti(x, names)
}
\arguments{
\item{x}{a list of \code{cptFor} objects}

\item{names}{a list of the names of the \code{cptFor} object being plotted}
}
\value{
ggplot
}
\description{
Plot a list of \code{cptFor} objects
}
