% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{replacementGeneric}
\alias{replacementGeneric}
\alias{errors<-}
\alias{m<-}
\alias{detector<-}
\alias{gamma<-}
\alias{errorsVar<-}
\alias{errors2Var<-}
\alias{cusum<-}
\alias{cusum2<-}
\alias{alpha<-}
\alias{critValue<-}
\alias{threshold<-}
\alias{threshold2<-}
\alias{tau<-}
\alias{tau2<-}
\alias{updateStats<-}
\alias{forecastErrorType<-}
\alias{updateStats2<-}
\title{Replacement Functions - Generic}
\usage{
errors(x) <- value

m(x) <- value

detector(x) <- value

gamma(x) <- value

errorsVar(x) <- value

errors2Var(x) <- value

cusum(x) <- value

cusum2(x) <- value

alpha(x) <- value

critValue(x) <- value

threshold(x) <- value

threshold2(x) <- value

tau(x) <- value

tau2(x) <- value

updateStats(x) <- value

forecastErrorType(x) <- value

updateStats2(x) <- value
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptFor}}}

\item{value}{value}
}
\description{
Replacement Functions - Generic
}
