% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{replacementMethod}
\alias{replacementMethod}
\alias{errors<-,cptFor-method}
\alias{m<-,cptFor-method}
\alias{detector<-,cptFor-method}
\alias{gamma<-,cptFor-method}
\alias{errorsVar<-,cptFor-method}
\alias{errors2Var<-,cptFor-method}
\alias{cusum<-,cptFor-method}
\alias{cusum2<-,cptFor-method}
\alias{alpha<-,cptFor-method}
\alias{threshold<-,cptFor-method}
\alias{threshold2<-,cptFor-method}
\alias{tau<-,cptFor-method}
\alias{tau2<-,cptFor-method}
\title{Replacement Functions - Method}
\usage{
\S4method{errors}{cptFor}(x) <- value

\S4method{m}{cptFor}(x) <- value

\S4method{detector}{cptFor}(x) <- value

\S4method{gamma}{cptFor}(x) <- value

\S4method{errorsVar}{cptFor}(x) <- value

\S4method{errors2Var}{cptFor}(x) <- value

\S4method{cusum}{cptFor}(x) <- value

\S4method{cusum2}{cptFor}(x) <- value

\S4method{alpha}{cptFor}(x) <- value

\S4method{threshold}{cptFor}(x) <- value

\S4method{threshold2}{cptFor}(x) <- value

\S4method{tau}{cptFor}(x) <- value

\S4method{tau2}{cptFor}(x) <- value
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptFor}}}

\item{value}{value}
}
\description{
Replacement Functions - Method
}
