% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{retrievalGeneric}
\alias{retrievalGeneric}
\alias{errors}
\alias{m}
\alias{detector}
\alias{errorsVar}
\alias{errors2Var}
\alias{cusum}
\alias{cusum2}
\alias{alpha}
\alias{critValue}
\alias{threshold}
\alias{threshold2}
\alias{tau}
\alias{tau2}
\alias{updateStats}
\alias{forecastErrorType}
\alias{updateStats2}
\title{Retrieval Functions - Generic}
\usage{
errors(x)

m(x)

detector(x)

errorsVar(x)

errors2Var(x)

cusum(x)

cusum2(x)

alpha(x)

critValue(x)

threshold(x)

threshold2(x)

tau(x)

tau2(x)

updateStats(x)

forecastErrorType(x)

updateStats2(x)
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptFor}}}
}
\description{
Retrieval Functions - Generic
}
