% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{retrievalMethod}
\alias{retrievalMethod}
\alias{errors,cptFor-method}
\alias{m,cptFor-method}
\alias{detector,cptFor-method}
\alias{gamma,cptFor-method}
\alias{errorsVar,cptFor-method}
\alias{errors2Var,cptFor-method}
\alias{cusum,cptFor-method}
\alias{cusum2,cptFor-method}
\alias{alpha,cptFor-method}
\alias{threshold,cptFor-method}
\alias{threshold2,cptFor-method}
\alias{tau,cptFor-method}
\alias{tau2,cptFor-method}
\title{Retrieval Functions - Method}
\usage{
\S4method{errors}{cptFor}(x)

\S4method{m}{cptFor}(x)

\S4method{detector}{cptFor}(x)

\S4method{gamma}{cptFor}(x)

\S4method{errorsVar}{cptFor}(x)

\S4method{errors2Var}{cptFor}(x)

\S4method{cusum}{cptFor}(x)

\S4method{cusum2}{cptFor}(x)

\S4method{alpha}{cptFor}(x)

\S4method{threshold}{cptFor}(x)

\S4method{threshold2}{cptFor}(x)

\S4method{tau}{cptFor}(x)

\S4method{tau2}{cptFor}(x)
}
\arguments{
\item{x}{object of class \code{\linkS4class{cptFor}}}
}
\description{
Retrieval Functions - Method
}
