% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateForecast.R
\name{updateForecast}
\alias{updateForecast}
\title{Update Forecast Errors}
\usage{
updateForecast(newErrors, model)
}
\arguments{
\item{newErrors}{numeric vector. New Forecast errors to perform changepoint analysis upon..}

\item{model}{\code{cptFor} object. Output of call to function \code{cptForecast} where initial forecast error
were analysed.}
}
\value{
An object of class \code{cptFor}
}
\description{
This function updates a \code{cptFor} object with new forecast errors. This allows for
the online implementation of the method as new forecast errors become available.
}
