% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateForecast.R
\name{updateForecastErrorChecks}
\alias{updateForecastErrorChecks}
\title{Error checking - updateForecast}
\usage{
updateForecastErrorChecks(newErrors = newErrors, model = model)
}
\arguments{
\item{newErrors}{numeric vector. New Forecast errors to perform changepoint analysis upon.}

\item{model}{\code{\linkS4class{cptFor}} object. Output of call to function \code{cptForecast} where
initial forecast errors were analysed.}
}
\description{
Performs error checking on arguments given to updateForecast
}
