% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptSeqCUSUM.R
\name{weightFun}
\alias{weightFun}
\title{CUSUM Weight Function}
\usage{
weightFun(m, k, gamma = 0)
}
\arguments{
\item{m}{Length of training period}

\item{k}{Time location}

\item{gamma}{tuning parameter}
}
\value{
numeric weight
}
\description{
CUSUM Weight Function
}
