% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\docType{class}
\name{cptFor-class}
\alias{cptFor-class}
\alias{cptFor}
\alias{plot,cptFor,ANY-method}
\alias{plot,cptFor-method}
\title{S4 class for a forecast changepoint object}
\usage{
\S4method{plot}{cptFor,ANY}(x)
}
\arguments{
\item{x}{an object of class \code{cptFor}}
}
\value{
An object of class \code{cptFor}.
}
\description{
List of the basic method
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot of \code{cptFor} object
}}

\section{Slots}{

\describe{
\item{\code{errors}}{ts or numeric vector. Forecast errors, one-step-ahead}

\item{\code{m}}{numeric. Length of training data}

\item{\code{X}}{ts or numeric vector. Raw time series}

\item{\code{model}}{ANY. Summary of model used for forecasting}

\item{\code{detector}}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}

\item{\code{gamma}}{numeric. Tuning parameter within detector}

\item{\code{errorsVar}}{numeric. Variance of forecast errors in training data}

\item{\code{errors2Var}}{numeric. Variance of squares of forecast errors}

\item{\code{cusum}}{numeric. CUSUM values of forecast errors}

\item{\code{cusum2}}{numeric. CUSUM values of squares of forecast errors}

\item{\code{alpha}}{numeric. Type=1 error}

\item{\code{threshold}}{numeric. threshold values for CUSUM values of forecast errors}

\item{\code{threshold2}}{numeric. threshold values for CUSUM values of squares of forecast errors}

\item{\code{tau}}{numeric. Time point when changepoint is flagged based upon forecast errors}

\item{\code{tau2}}{numeric. Time point when changepoint is flagged based upon squares of forecast errors}
}}

