% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\docType{class}
\name{cptFor-class}
\alias{cptFor-class}
\alias{cptFor}
\title{S4 class for a forecast changepoint object}
\value{
An object of class \code{cptFor}.
}
\description{
An S4 class that contains the results of the analysis of sequential changepoint detection.
\itemize{
\item For a list of methods that can be performed on a \code{cptFor} object see \code{\link{cptFor-methods}}.
\item For information about the retrieval functions for the slots see \code{\link{retrievalMethod}}.
}
Note with many of the slots a postfix 2 indicates it is related to the analysis on the squared
forecast errors. For example \code{tau2} is the changepoint location using the squared forecast errors.
}
\section{Slots}{

\describe{
\item{\code{errors}}{numeric vector. Forecast errors, one-step-ahead}

\item{\code{m}}{numeric. Length of training data}

\item{\code{detector}}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}

\item{\code{forecastErrorType}}{character. Type of changes to look for. Choice of
\itemize{
\item{"Both": }{Analysis is performed on both the raw and squared forecast errors,}
\item{"Raw": }{Analysis is only performed on the raw forecast errors. Only first
order changes are reliably detected.}
\item{"Squared:" }{Analysis is only performed on the centred squared forecast errors.}
}}

\item{\code{gamma}}{numeric. Tuning parameter within detector}

\item{\code{errorsVar}}{numeric. Variance of forecast errors in training data}

\item{\code{errors2Var}}{numeric. Variance of squares of forecast errors}

\item{\code{cusum}}{numeric. CUSUM values of forecast errors}

\item{\code{cusum2}}{numeric. CUSUM values of squares of forecast errors}

\item{\code{alpha}}{numeric. Type-1 error}

\item{\code{critValue}}{numeric. Critical value used in the threshold to maintain type-1 error rate}

\item{\code{threshold}}{numeric. threshold values for CUSUM values of forecast errors}

\item{\code{threshold2}}{numeric. threshold values for CUSUM values of squares of forecast errors}

\item{\code{tau}}{numeric. Time point when changepoint is flagged based upon forecast errors}

\item{\code{tau2}}{numeric. Time point when changepoint is flagged based upon squares of forecast errors}

\item{\code{updateStats}}{numeric vector. Update statistics needed for use with \code{updateForecast}}

\item{\code{updateStats2}}{numeric vector. Update statistics needed for use with \code{updateForecast}}

\item{\code{version}}{character. Version of changepoint.forecast version}

\item{\code{date}}{character. Date object was created/updated}
}}

\examples{
ans = cptForecast(c(stats::rnorm(400), stats::rnorm(100,3)), m=300)
summary(ans)
show(ans)
plot(ans)
}
