% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptForClass.R
\name{cptFor-methods}
\alias{cptFor-methods}
\alias{summary,cptFor-method}
\alias{show,cptFor-method}
\title{cptFor Methods}
\usage{
\S4method{summary}{cptFor}(object)

\S4method{show}{cptFor}(object)
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{cptFor}}}
}
\description{
Methods for objects with S4 class \code{\linkS4class{cptFor}}
}
\section{Functions}{
\itemize{
\item \code{summary,cptFor-method}: Summary of \code{cptFor} object

\item \code{show,cptFor-method}: Show \code{cptFor} object
}}

