% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critValLookup.R
\name{critValLookup}
\alias{critValLookup}
\title{Lookup Critical Value}
\usage{
critValLookup(gamma = 0, alpha = 0.05, detector = "PageCUSUM")
}
\arguments{
\item{gamma}{tuning parameter}

\item{alpha}{type 1 error}

\item{detector}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}
}
\value{
numeric.
}
\description{
Lookup pre-simulated critical value of asymptotic distribution of chosen CUSUM detector
}
\examples{
critValLookup(gamma=0,
              alpha=0.05,
              detector='PageCUSUM')
}
