% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCritVal.R
\name{simCritVal}
\alias{simCritVal}
\title{Simulate Critical Value}
\usage{
simCritVal(
  samples = 1000,
  alpha = 0.05,
  detector = "PageCUSUM",
  gamma = 0,
  npts = 500,
  progressBar = TRUE
)
}
\arguments{
\item{samples}{Number of samples to simulate from asymptotic distribution}

\item{alpha}{Type 1 error}

\item{detector}{character. Type of changepoint detector to use. Choice of
\itemize{
\item{"PageCUSUM": }{Page's CUSUM detector for 2-sided alternative hypothesis}
\item{"PageCUSUM1": }{Page's CUSUM detector for 1-sided alternative hypothesis}
\item{"CUSUM": }{Original CUSUM detector for 2-sided alternative hypothesis}
\item{"CUSUM1": }{Original CUSUM detector for 1-sided alternative hypothesis}
}}

\item{gamma}{tuning parameter in the weight function}

\item{npts}{Number of points in discretization of Weiner process}

\item{progressBar}{Logical. Show progress bar}
}
\value{
numeric critical value
}
\description{
Simulates a critical value from the asymptotic distribution of the chosen CUSUM detector under the no change scenario.
}
\examples{
ans = simCritVal(samples=100, npts=20)
}
