% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateForecast.R
\name{updateForecast}
\alias{updateForecast}
\title{Update Analysis of Forecast Errors}
\usage{
updateForecast(newErrors, model)
}
\arguments{
\item{newErrors}{numeric vector. New Forecast errors to perform changepoint analysis upon.}

\item{model}{\code{\linkS4class{cptFor}} object. Output of call to function \code{cptForecast} where
initial forecast errors were analysed.}
}
\value{
An object of class \code{cptFor}
}
\description{
This function updates the analysis in a \code{cptFor} object with new forecast errors. This allows for
the online implementation of the method as new forecast errors become available.
}
\details{
This function can be embedded within a framework to allow for the sequential analysis of forecast
errors as new data becomes available. Say you receive a new forecast error each day then this
function can be used to update the analysis and then a flag can be created to signal when a
changepoint has been detected by monitoring the \code{tau} and \code{tau2} slots.
}
\examples{
ans = cptForecast(stats::rnorm(500), m=300)
ans = updateForecast(stats::rnorm(10, 3), model=ans)
summary(ans)
plot(ans)
}
